third_party/blink/web_tests/platform/webview/external/wpt/webvtt/api/idlharness.window-expected.txt

This is a testharness.js-based test.
Found 30 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] VTTCue interface: attribute region
  assert_true: The prototype object must have a property "region" expected true got false
[FAIL] VTTCue interface: attribute lineAlign
  assert_true: The prototype object must have a property "lineAlign" expected true got false
[FAIL] VTTCue interface: attribute positionAlign
  assert_true: The prototype object must have a property "positionAlign" expected true got false
[FAIL] VTTCue interface: new VTTCue(0, 0, "") must inherit property "region" with the proper type
  assert_inherits: property "region" not found in prototype chain
[FAIL] VTTCue interface: new VTTCue(0, 0, "") must inherit property "lineAlign" with the proper type
  assert_inherits: property "lineAlign" not found in prototype chain
[FAIL] VTTCue interface: new VTTCue(0, 0, "") must inherit property "positionAlign" with the proper type
  assert_inherits: property "positionAlign" not found in prototype chain
[FAIL] VTTRegion interface: existence and properties of interface object
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface object length
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface object name
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: existence and properties of interface prototype object
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: existence and properties of interface prototype object's "constructor" property
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: existence and properties of interface prototype object's @@unscopables property
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute id
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute width
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute lines
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute regionAnchorX
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute regionAnchorY
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute viewportAnchorX
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute viewportAnchorY
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion interface: attribute scroll
  assert_own_property: self does not have own property "VTTRegion" expected property "VTTRegion" missing
[FAIL] VTTRegion must be primary interface of new VTTRegion()
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] Stringification of new VTTRegion()
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "id" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "width" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "lines" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "regionAnchorX" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "regionAnchorY" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "viewportAnchorX" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "viewportAnchorY" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
[FAIL] VTTRegion interface: new VTTRegion() must inherit property "scroll" with the proper type
  assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: VTTRegion is not defined"
Harness: the test ran to completion.
